+++ the-shop-app/assets/favicon.png
