+++ the-shop-app/assets/icon.png
