+++ the-shop-app/assets/splash.png
