--- the-shop-app/components/StyledText.tsx
@@ -1,7 +1,0 @@
-import * as React from 'react';
-
-import { Text, TextProps } from './Themed';
-
-export function MonoText(props: TextProps) {
-  return <Text {...props} style={[props.style, { fontFamily: 'space-mono' }]} />;
-}
