--- the-shop-app/hooks/useColorScheme.web.ts
@@ -1,5 +1,0 @@
-// useColorScheme from react-native does not support web currently. You can replace
-// this with react-native-appearance if you would like theme support on web.
-export default function useColorScheme() {
-  return 'light';
-}
