--- the-shop-app/navigation/LinkingConfiguration.ts
@@ -1,24 +1,0 @@
-import * as Linking from 'expo-linking';
-
-export default {
-  prefixes: [Linking.makeUrl('/')],
-  config: {
-    screens: {
-      Root: {
-        screens: {
-          TabOne: {
-            screens: {
-              TabOneScreen: 'one',
-            },
-          },
-          TabTwo: {
-            screens: {
-              TabTwoScreen: 'two',
-            },
-          },
-        },
-      },
-      NotFound: '*',
-    },
-  },
-};
